Modifs Vue

+++ Shopping/src/Vue/VueBestSellers.java
@@ -5,7 +5,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.util.ArrayList;
-
+//classe qui affiche les produits les plus achetes
 public class VueBestSellers extends JPanel {
     private final Color backgroundColor = new Color(220, 223, 197);
 
@@ -99,7 +99,7 @@ public class VueBestSellers extends JPanel {
                 JPanel productPanel = ProductPanelFactory.createProductPanel(
                         article.getNomArticle(),         // Nom du produit
                         article.getCategorieArticle(),   // Description ou catégorie
-                        article.getPrixArticle(),        // Prix
+                        (Double)(double)article.getPrixArticle(),        // Prix
                         article.getNomArticle()        // Nom du porduit
                 );
